refactor(openmemory): extract synaptic code into mcp_synaptic.py
Move all synaptic memory logic (singleton, add/search hooks, 3 MCP
tools) out of mcp_server.py into a dedicated mcp_synaptic.py module.
Reduces fork diff in upstream file from ~45 interleaved lines to 4
stable one-line delegations, minimizing cherry-pick conflicts.

Also adds synaptic_memory/pyproject.toml so the package is
pip-installable without the PYTHONPATH hack.
